fixed to precision level of long double

+++ q3/libpoisson.c
@@ -2,10 +2,10 @@
 #include <math.h>   // pow()
 
 long double poisson (long double lambda, long double k){
-    return (pow(lambda, k)/factorial(k)) * expf(-(int)lambda);
+    return ((long double)(pow(lambda, k))/factorial(k)) * (long double)(expf(-lambda));
 }
 
-long factorial(int n){
+long double factorial(long double n){
     if (n == 0 || n == 1)
         return 1;
     return factorial(n-1) * n;
